# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hivemind Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lattice/run:
    post:
      operationId: runLattice
      tags: [lattice]
      summary: Run Hivemind Predictive Lattice (HPL-HPA v2) for a symbol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunLatticeBody"
      responses:
        "200":
          description: Full lattice result with belief tokens, debate rounds, SHAP, and causal narrative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatticeResult"
  /lattice/challenge:
    post:
      operationId: latticeChallenge
      tags: [lattice]
      summary: Challenge a lattice agent with a new argument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LatticeChallengeBody"
      responses:
        "200":
          description: Agent response and updated probability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatticeChallengeResponse"
...
  /lattice/regime:
    get:
      operationId: getMarketRegime
      tags: [lattice]
      summary: Detect current volatility regime for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Regime status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegimeStatus"
...
components:
  schemas:
...
    RunLatticeBody:
      type: object
      properties:
        symbol:
          type: string
        timeframe:
          type: string
          default: "7d"
      required: [symbol]
    LatticeResult:
      type: object
      properties:
        runId:
          type: string
        symbol:
          type: string
        timeframe:
          type: string
        regime:
          type: string
          enum: [calm, volatile, crisis]
        regimeScore:
          type: number
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/BeliefToken"
        debateRounds:
          type: array
          items:
            $ref: "#/components/schemas/DebateRound"
        shap:
          $ref: "#/components/schemas/ShapBreakdown"
        finalPrediction:
          $ref: "#/components/schemas/LatticePrediction"
        causalNarrative:
          type: string
        minorityReport:
          type: string
          nullable: true
        agentConsensus:
          type: number
        polymarketIntel:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/PolymarketIntelItem"
      required:
        [
          runId,
          symbol,
          timeframe,
          regime,
          regimeScore,
          tokens,
          debateRounds,
          shap,
          finalPrediction,
          causalNarrative,
          agentConsensus,
        ]
    PolymarketIntelItem:
      type: object
      properties:
        headline:
          type: string
          description: Short breaking-news style headline about the event
        question:
          type: string
        yesPrice:
          type: number
        oddsShift:
          type: number
          nullable: true
        marketImpact:
          type: string
        category:
          type: string
        volume:
          type: number
      required: [headline, question, yesPrice, marketImpact, category, volume]
    BeliefToken:
      type: object
      properties:
        id:
          type: string
        agentType:
          type: string
        round:
          type: integer
        hypothesis:
          type: string
          enum: [bullish, bearish, neutral]
        probability:
          type: number
        confidence:
          type: number
        rationale:
          type: array
          items:
            type: string
        shapHive:
          type: number
        shapAi:
          type: number
        shapGeo:
          type: number
        liquidityScore:
          type: number
        parentIds:
          type: array
          items:
            type: string
      required:
        [
          id,
          agentType,
          round,
          hypothesis,
          probability,
          confidence,
          rationale,
          shapHive,
          shapAi,
          shapGeo,
          liquidityScore,
          parentIds,
        ]
    DebateRound:
      type: object
      properties:
        round:
          type: integer
        agentType:
          type: string
        challenge:
          type: string
        adjustment:
          type: number
        accepted:
          type: boolean
      required: [round, agentType, challenge, adjustment, accepted]
    ShapBreakdown:
      type: object
      properties:
        hive:
          type: number
        ai:
          type: number
        geo:
          type: number
      required: [hive, ai, geo]
    LatticePrediction:
      type: object
      properties:
        direction:
          type: string
          enum: [bullish, bearish, neutral]
        targetPrice:
          type: number
        confidence:
          type: number
        hivemindScore:
          type: number
      required: [direction, targetPrice, confidence, hivemindScore]
...
    LatticeChallengeBody:
      type: object
      properties:
        agentType:
          type: string
        challenge:
          type: string
        symbol:
          type: string
        currentProbability:
          type: number
      required: [agentType, challenge, symbol, currentProbability]
    LatticeChallengeResponse:
      type: object
      properties:
        agentType:
          type: string
        response:
          type: string
        adjustment:
          type: number
        newProbability:
          type: number
      required: [agentType, response, adjustment, newProbability]
    RegimeStatus:
      type: object
      properties:
        symbol:
          type: string
        regime:
          type: string
          enum: [calm, volatile, crisis]
        regimeScore:
          type: number
        volatility:
          type: number
        description:
          type: string
      required: [symbol, regime, regimeScore, volatility, description]